Fix career RPC type casts + Career Mode card with New/Load buttons
- Cast tier param to SMALLINT in all rpc_generate_career_opponents calls
  (fixes 'function does not exist' error on career creation)
- Replace Career Mode GameModeCard with custom card having two buttons:
  'New Career' -> /app/career/start
  'Load Career' -> /app/career (shows save picker or auto-loads)

+++ supabase/migrations/20260304_career_bracket_rpcs.sql
@@ -102,7 +102,7 @@ BEGIN
   -- If not enough opponents, generate more
   IF v_count < (v_bracket_size - 1) THEN
     PERFORM rpc_generate_career_opponents(
-      p_career_id, v_career.tier,
+      p_career_id, v_career.tier::SMALLINT,
       (v_bracket_size - 1 - v_count),
       v_career.career_seed + v_career.season * 1000 + v_career.week * 10
     );
